+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/sql/QueryBuilder.java
@@ -899,6 +899,10 @@ public class QueryBuilder {
             throw new IllegalArgumentException();
         }
 
+        if (values.isEmpty()) {
+            return this;
+        }
+
         prettyPrinter.write("WITH");
         prettyPrinter.begin(2);
         prettyPrinter.newline(0);