+++ RELEASES.md
@@ -1,4 +1,7 @@
 See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
+`2.0.5-SNAPSHOT`
+* prov-template-library: cleaned up pom.xml
+* prov-template-compiler: allowing for @sql.type in NameDescriptor
 
 `2.0.4`
 * prov-explanations: restructured explanation system, introduced interfaces separating components
@@ -358,6 +361,6 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * Refactored converters
 * Deprecated old traversal classes. They will be deleted in next release.
 
-`0.3.0`
+`0.0.3`
 
 * Version included in prov-translator service, launched on Dec 17th